requirements.txt
djangorestframework==3.9.4
django-cors-headers==2.4.0
